[Rename] Sort Out Board DTO
