+++ src/main/java/com/example/boardserver/board/controller/BoardController.java
@@ -3,7 +3,12 @@ package com.example.boardserver.board.controller;
 import com.example.boardserver.auth.dto.CustomUserDetails;
 import com.example.boardserver.board.converter.BoardConverter;
 import com.example.boardserver.board.domain.BoardLikeId;
-import com.example.boardserver.board.dto.*;
+import com.example.boardserver.board.dto.boardDTO.BoardRequestDTO;
+import com.example.boardserver.board.dto.boardDTO.BoardResponseDTO;
+import com.example.boardserver.board.dto.boardDTO.BoardResponseDetailDTO;
+import com.example.boardserver.board.dto.commentDTO.CommentListResponseDTO;
+import com.example.boardserver.board.dto.commentDTO.CommentRequestDTO;
+import com.example.boardserver.board.dto.commentDTO.CommentResponseDTO;
 import com.example.boardserver.board.service.commentService.CommentCommandService;
 import com.example.boardserver.board.service.commentService.CommentQueryService;
 import com.example.boardserver.board.service.boardLikeService.BoardLikeCommandService;
