+++ src/main/java/com/example/boardserver/board/service/boardService/BoardQueryServiceImpl.java
@@ -2,7 +2,7 @@ package com.example.boardserver.board.service.boardService;
 
 import com.example.boardserver.board.converter.BoardConverter;
 import com.example.boardserver.board.domain.Board;
-import com.example.boardserver.board.dto.BoardResponseDetailDTO;
+import com.example.boardserver.board.dto.boardDTO.BoardResponseDetailDTO;
 import com.example.boardserver.board.repository.boardRepository.BoardRepository;
 import com.example.boardserver.common.code.status.ErrorStatus;
 import com.example.boardserver.exception.handler.BoardHandler;
