+++ src/main/java/com/example/boardserver/board/service/commentService/CommentCommandService.java
@@ -1,7 +1,7 @@
 package com.example.boardserver.board.service.commentService;
 
-import com.example.boardserver.board.dto.CommentRequestDTO;
-import com.example.boardserver.board.dto.CommentResponseDTO;
+import com.example.boardserver.board.dto.commentDTO.CommentRequestDTO;
+import com.example.boardserver.board.dto.commentDTO.CommentResponseDTO;
 
 public interface CommentCommandService {
