+++ src/main/java/com/example/boardserver/board/service/commentService/CommentCommandServiceImpl.java
@@ -3,8 +3,8 @@ package com.example.boardserver.board.service.commentService;
 import com.example.boardserver.board.converter.CommentConverter;
 import com.example.boardserver.board.domain.Board;
 import com.example.boardserver.board.domain.Comment;
-import com.example.boardserver.board.dto.CommentRequestDTO;
-import com.example.boardserver.board.dto.CommentResponseDTO;
+import com.example.boardserver.board.dto.commentDTO.CommentRequestDTO;
+import com.example.boardserver.board.dto.commentDTO.CommentResponseDTO;
 import com.example.boardserver.board.repository.commentRepository.CommentRepository;
 import com.example.boardserver.board.repository.boardRepository.BoardRepository;
 import com.example.boardserver.common.code.status.ErrorStatus;
